+++ internal/bus/bus.go
@@ -2,10 +2,10 @@ package bus
 
 import (
 	"fmt"
+	"github.com/c3os-io/c3os/sdk/bus"
 	"os"
 	"os/exec"
 
-	"github.com/c3os-io/c3os/pkg/bus"
 	"github.com/mudler/go-pluggable"
 )
 
@@ -15,6 +15,7 @@ var Manager = &Bus{
 		[]pluggable.EventType{
 			bus.EventBootstrap,
 			bus.EventChallenge,
+			bus.EventBoot,
 			bus.EventInstall,
 		},
 	),
