+++ internal/provider/install.go
@@ -3,8 +3,8 @@ package provider
 import (
 	"context"
 	"encoding/json"
+	"github.com/c3os-io/c3os/sdk/bus"
 
-	"github.com/c3os-io/c3os/pkg/bus"
 	"github.com/mudler/go-nodepair"
 	"github.com/mudler/go-pluggable"
 )
